tests/requirements.txt
pytest
pytest-cov
python-dotenv
psycopg[binary]
redis
